+++ Projects/UOContent/Mobiles/AI/BaseAI/BaseAI.cs
@@ -623,139 +623,6 @@ public abstract partial class BaseAI
         }
     }
 
-    private void HandleNoOrder()
-    {
-        m_Mobile.ControlTarget = null;
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-    }
-    
-    private void HandleTransferOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-        
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.ControlTarget = null;
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-        m_Mobile.PlaySound(m_Mobile.GetIdleSound());
-        _lastCommandIssuer = null;
-    }
-    
-    private void HandleGuardOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = true;
-        m_Mobile.PlaySound(m_Mobile.GetAttackSound());
-        m_Mobile.ControlMaster.SendLocalizedMessage(1049671, m_Mobile.Name);
-        // 1049671: ~1_NAME~ is now guarding you.
-        _lastCommandIssuer = null;
-    }
-    
-    private void HandleAttackOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = true;
-        m_Mobile.PlaySound(m_Mobile.GetAttackSound());
-        _lastCommandIssuer = null;
-    }
-    
-    private void HandleFollowOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-        m_Mobile.PlaySound(m_Mobile.GetIdleSound());
-        _lastCommandIssuer = null;
-    }
-
-    private void HandleStayOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-        m_Mobile.PlaySound(m_Mobile.GetIdleSound());
-        m_Mobile.Home = m_Mobile.Location;
-        _lastCommandIssuer = null;
-    }
-
-    private void HandleStopOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.ControlTarget = null;
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-        m_Mobile.PlaySound(m_Mobile.GetIdleSound());
-        _lastCommandIssuer = null;
-    }
-
-    private void HandleReleaseOrder()
-    {
-        if (m_Mobile.ControlMaster?.Alive != true)
-        {
-            return;
-        }
-
-        _lastCommandIssuer?.RevealingAction();
-        m_Mobile.ControlTarget = null;
-        m_Mobile.FocusMob = null;
-        m_Mobile.Warmode = false;
-        m_Mobile.Combatant = null;
-        m_Mobile.PlaySound(m_Mobile.GetIdleSound());
-        m_Mobile.BondingBegin = DateTime.MinValue;
-        m_Mobile.OwnerAbandonTime = DateTime.MinValue;
-        m_Mobile.IsBonded = false;
-        m_Mobile.SetControlMaster(null);
-        _lastCommandIssuer = null;
-    }
-
-    public virtual void HandleRenameOrder()
-    {
-        if (m_Mobile.Summoned)
-        {
-            m_Mobile.ControlMaster?.SendMessage("You cannot rename a summoned creature.");
-        }
-        else
-        {
-            m_Mobile.ControlMaster?.SendMessage("Change name on pet health bar.");
-        }
-    }
-
     public virtual bool CheckHerding()
     {
         var target = m_Mobile.TargetLocation;
